Moved block handling methods into a separate file because they were using calls to stdlib and math that were interfearing with pyopencl kernel compilation (need to figure out why this is happening).
